Replace public js files in components folder and access these with main.js file, which we will bundle
