+++ package.json
@@ -8,7 +8,7 @@
     "build-sass": "node-sass --include-path scss sass/main.scss public/css/main.css",
     "watch-sass": "node-sass -w --include-path scss sass/main.scss public/css/main.css",
     "build-js": "browserify public/js/*.js -o public/dist/bundle.js",
-    "watch-js": "watchify public/js/*.js -o public/dist/bundle.js -dv",
+    "watch-js": "watchify public/js/main.js -o public/dist/bundle.js -dv",
     "build": "npm run build-sass | npm run build-js"
   },
   "repository": {
