+++ public/js/components/image-detail.js
