+++ public/js/components/new-story-submit.js
