+++ public/js/main.js
@@ -1,0 +1,5 @@
+const map = require('./components/map.js');
+
+(function () {
+  map.init();
+}) ();
